@@ -1,6 +1,6 @@
 {
   "name": "@senzing/sdk-components-ng",
-  "version": "2.0.1",
+  "version": "2.1.0",
   "private": false,
   "keywords" :["Angular","Library","Senzing"],
   "license" : "SEE LICENSE IN LICENSE",
@@ -19,9 +19,9 @@
     "@types/jspdf": "^1.3.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/core": "~7.1.0",
+    "@angular/cdk": "^10.1.3",
+    "@angular/common": "~10.0.9",
+    "@angular/core": "~10.0.9",
     "@senzing/sdk-graph-components": "^2.0.0"
   }
 }
